Rename kutils to signals
also moved out methods that do not "belong" in signals.py
