+++ kwave/utils/__init__.py
@@ -1,14 +1,14 @@
-from kwave.utils.dotdictionary import *
 from kwave.utils.checks import *
 from kwave.utils.colormap import *
 from kwave.utils.conversion import *
 from kwave.utils.data import *
+from kwave.utils.dotdictionary import *
 from kwave.utils.filters import *
 from kwave.utils.interp import *
 from kwave.utils.io import *
-from kwave.utils.kutils import *
 from kwave.utils.mapgen import *
+from kwave.utils.math import *
 from kwave.utils.matrix import *
 from kwave.utils.pml import *
+from kwave.utils.signals import *
 from kwave.utils.tictoc import *
-from kwave.utils.math import *
