+++ tests/test_filterutils.py
@@ -1,8 +1,10 @@
-import numpy as np
-from kwave.utils.filters import gaussian, gaussian_filter, create_cw_signals, fwhm, envelope_detection, sharpness, \
-    smooth
-from kwave.utils.kutils import get_win
 from math import pi
+
+import numpy as np
+
+from kwave.utils.filters import gaussian, create_cw_signals, fwhm, envelope_detection, sharpness, \
+    smooth
+from kwave.utils.signals import get_win
 
 
 def test_gaussian():
