+++ tests/test_us_bmode_linear_transducer.py
@@ -6,18 +6,14 @@
     structure. It builds on the Defining An Ultrasound Transducer and
     Simulating Ultrasound Beam Patterns examples.
 """
+from tempfile import gettempdir
+
 # noinspection PyUnresolvedReferences
 import setup_test
-import os
-from tempfile import gettempdir
+from kwave.kmedium import kWaveMedium
 from kwave.kspaceFirstOrder3D import kspaceFirstOrder3DC
-from kwave.utils.kutils import tone_burst
-from kwave.utils.mapgen import make_ball
-from kwave.utils import dotdict
 from kwave.ktransducer import *
 from tests.diff_utils import compare_against_ref
-from kwave.kmedium import kWaveMedium
-
 
 """
     Differences compared to original example:
