+++ tests/test_us_defining_transducer.py
@@ -6,17 +6,14 @@
     structure. It builds on the Defining An Ultrasound Transducer and
     Simulating Ultrasound Beam Patterns examples.
 """
+from tempfile import gettempdir
+
 # noinspection PyUnresolvedReferences
 import setup_test
-import os
-from tempfile import gettempdir
-from kwave.kgrid import kWaveGrid
+from kwave.kmedium import kWaveMedium
 from kwave.kspaceFirstOrder3D import kspaceFirstOrder3DC
-from kwave.utils.kutils import tone_burst
-from kwave.utils import dotdict
 from kwave.ktransducer import *
 from tests.diff_utils import compare_against_ref
-from kwave.kmedium import kWaveMedium
 
 
 def test_us_defining_transducer():
@@ -82,7 +79,7 @@ def test_us_defining_transducer():
     tone_burst_cycles = 5
 
     # create the input signal using tone_burst
-    input_signal = tone_burst(1 / kgrid.dt, tone_burst_freq, tone_burst_cycles);
+    input_signal = tone_burst(1 / kgrid.dt, tone_burst_freq, tone_burst_cycles)
 
     # scale the source magnitude by the source_strength divided by the
     # impedance (the source is assigned to the particle velocity)
