+++ tests/test_utils.py
@@ -1,15 +1,15 @@
-from kwave.utils.checks import num_dim
-from kwave.utils.mapgen import hounsfield2density, fit_power_law_params, power_law_kramers_kronig, make_cart_circle, make_cart_sphere
-from kwave.utils.conversion import db2neper, neper2db
-from kwave.utils.kutils import tone_burst, add_noise, gradient_spect
-from kwave.utils.interp import get_bli
-from kwave.utils.filters import extract_amp_phase, spect, apply_filter
-from kwave.utils.matrix import gradient_fd, resize
-from kwave.utils.io import load_image
 import numpy as np
-from utils import *
 import pytest
 from phantominator import shepp_logan
+
+from kwave.utils.checks import num_dim
+from kwave.utils.conversion import db2neper, neper2db
+from kwave.utils.filters import extract_amp_phase, spect, apply_filter
+from kwave.utils.interp import get_bli
+from kwave.utils.mapgen import hounsfield2density, fit_power_law_params, power_law_kramers_kronig, make_cart_circle, \
+    make_cart_sphere
+from kwave.utils.matrix import gradient_fd, resize
+from kwave.utils.signals import tone_burst, add_noise, gradient_spect
 
 input_signal = np.array([0., 0.00099663, 0.00646706, 0.01316044, 0.01851998,
                          0.02355139, 0.02753738, 0.02966112, 0.02907933, 0.02502059,
